split camera utils unit tests into their own file to make it easier to track down bugs in the future
