+++ tests/testAstrometry.py
@@ -25,7 +25,6 @@ from collections import OrderedDict
 import lsst.utils.tests as utilsTests
 from lsst.utils import getPackageDir
 
-import lsst.afw.geom as afwGeom
 from lsst.sims.utils import ObservationMetaData
 from lsst.sims.utils import getRotTelPos, raDecFromAltAz, calcObsDefaults, \
                             radiansFromArcsec, arcsecFromRadians, Site
@@ -34,13 +33,6 @@ from lsst.sims.coordUtils import applyPrecession, applyProperMotion
 from lsst.sims.coordUtils import appGeoFromICRS, observedFromAppGeo
 from lsst.sims.coordUtils import observedFromICRS, calculatePupilCoordinates
 from lsst.sims.coordUtils import refractionCoefficients, applyRefraction
-from lsst.sims.coordUtils import calculateFocalPlaneCoordinates
-from lsst.sims.coordUtils import findChipName, calculatePixelCoordinates
-from lsst.sims.coordUtils import raDecFromPupilCoordinates, raDecFromPixelCoordinates
-from lsst.sims.coordUtils import pupilCoordinatesFromPixelCoordinates
-import lsst.afw.cameraGeom.testUtils as camTestUtils
-
-from lsst.sims.coordUtils.utils import ReturnCamera
 
 def makeObservationMetaData():
     #create a cartoon ObservationMetaData object
@@ -345,216 +337,6 @@ class astrometryUnitTest(unittest.TestCase):
         dec = numpy.random.random_sample(10)*numpy.radians(1.0) + numpy.radians(dummy.unrefractedDec)
         test = calculatePupilCoordinates(ra, dec, obs_metadata=dummy, epoch=2000.0)
 
-    def testCameraCoordsExceptions(self):
-        """
-        Test to make sure that focal plane methods raise exceptions when coordinates are improperly
-        specified.
-        """
-
-        camera = camTestUtils.CameraWrapper().camera
-
-        #these are just values shown heuristically to give an actual chip name
-        ra = numpy.array(numpy.radians(self.obs_metadata.unrefractedRA) - numpy.array([1.01, 1.02])*numpy.radians(1.0/3600.0))
-        dec = numpy.array(numpy.radians(self.obs_metadata.unrefractedDec) - numpy.array([2.02, 2.01])*numpy.radians(1.0/3600.0))
-
-        ra, dec = observedFromICRS(ra, dec, obs_metadata=self.obs_metadata, epoch=2000.0)
-
-        xPupil = numpy.array([-0.000262243770, -0.00000234])
-        yPupil = numpy.array([0.000199467792, 0.000189334])
-
-        ##########test findChipName
-
-        name = findChipName(ra=ra, dec=dec,
-                            epoch=2000.0,
-                            obs_metadata=self.obs_metadata,
-                            camera=camera)
-
-        self.assertTrue(name[0] is not None)
-
-        name = findChipName(xPupil=xPupil, yPupil=yPupil,
-                            camera=camera)
-
-        self.assertTrue(name[0] is not None)
-
-        #test when specifying no coordinates
-        self.assertRaises(RuntimeError, findChipName)
-
-        #test when specifying both sets fo coordinates
-        self.assertRaises(RuntimeError, findChipName, xPupil=xPupil, yPupil=yPupil,
-                  ra=ra, dec=dec, camera=camera)
-
-        #test when failing to specify camera
-        self.assertRaises(RuntimeError, findChipName, ra=ra, dec=dec,
-                          obs_metadata=self.obs_metadata, epoch=2000.0)
-        self.assertRaises(RuntimeError, findChipName, xPupil=xPupil, yPupil=yPupil)
-
-        #test when failing to specify obs_metadata
-        self.assertRaises(RuntimeError, findChipName, ra=ra, dec=dec, epoch=2000.0,
-                          camera=camera)
-
-        #test when failing to specify epoch
-        self.assertRaises(RuntimeError, findChipName, ra=ra, dec=dec, camera=camera,
-                          obs_metadata=self.obs_metadata)
-
-        #test mismatches
-        self.assertRaises(RuntimeError, findChipName, ra=numpy.array([ra[0]]), dec=dec,
-                            epoch=2000.0,
-                            obs_metadata=self.obs_metadata,
-                            camera=camera)
-
-        self.assertRaises(RuntimeError, findChipName, ra=ra, dec=numpy.array([dec[0]]),
-                            epoch=2000.0,
-                            obs_metadata=self.obs_metadata,
-                            camera=camera)
-
-        self.assertRaises(RuntimeError, findChipName, xPupil=numpy.array([xPupil[0]]), yPupil=yPupil,
-                                        camera=camera)
-        self.assertRaises(RuntimeError, findChipName, xPupil=xPupil, yPupil=numpy.array([yPupil[0]]),
-                                        camera=camera)
-
-        #test lists
-        self.assertRaises(RuntimeError, findChipName, ra=list(ra), dec=dec,
-                            epoch=2000.0,
-                            obs_metadata=self.obs_metadata,
-                            camera=camera)
-
-        self.assertRaises(RuntimeError, findChipName, ra=ra, dec=list(dec),
-                            epoch=2000.0,
-                            obs_metadata=self.obs_metadata,
-                            camera=camera)
-
-        self.assertRaises(RuntimeError, findChipName, xPupil=list(xPupil), yPupil=yPupil,
-                                        camera=camera)
-        self.assertRaises(RuntimeError, findChipName, xPupil=xPupil, yPupil=list(yPupil),
-                                        camera=camera)
-
-
-        ##########test FocalPlaneCoordinates
-
-        #test that it actually runs
-        xx, yy = calculateFocalPlaneCoordinates(xPupil=xPupil, yPupil=yPupil, camera=camera)
-        xx, yy = calculateFocalPlaneCoordinates(ra=ra, dec=dec,
-                                                epoch=2000.0, obs_metadata=self.obs_metadata,
-                                                camera=camera)
-
-        #test without any coordinates
-        self.assertRaises(RuntimeError, calculateFocalPlaneCoordinates, camera=camera)
-
-        #test specifying both ra,dec and xPupil,yPupil
-        self.assertRaises(RuntimeError, calculateFocalPlaneCoordinates, ra=ra, dec=dec,
-                             xPupil=xPupil, yPupil=yPupil, camera=camera)
-
-        #test without camera
-        self.assertRaises(RuntimeError, calculateFocalPlaneCoordinates, xPupil=xPupil, yPupil=yPupil)
-        self.assertRaises(RuntimeError, calculateFocalPlaneCoordinates, ra=ra, dec=dec,
-                                        epoch=2000.0, obs_metadata=self.obs_metadata)
-
-        #test without epoch
-        self.assertRaises(RuntimeError, calculateFocalPlaneCoordinates, ra=ra, dec=dec,
-                                                obs_metadata=self.obs_metadata,
-                                                camera=camera)
-
-        #test without obs_metadata
-        self.assertRaises(RuntimeError, calculateFocalPlaneCoordinates, ra=ra, dec=dec,
-                                               epoch=2000.0,
-                                               camera=camera)
-
-        #test with lists
-        self.assertRaises(RuntimeError, calculateFocalPlaneCoordinates, xPupil=list(xPupil), yPupil=yPupil,
-                          camera=camera)
-        self.assertRaises(RuntimeError, calculateFocalPlaneCoordinates, xPupil=xPupil, yPupil=list(yPupil),
-                          camera=camera)
-
-        self.assertRaises(RuntimeError, calculateFocalPlaneCoordinates, ra=list(ra), dec=dec,
-                                        epoch=2000.0, camera=camera)
-        self.assertRaises(RuntimeError, calculateFocalPlaneCoordinates, ra=ra, dec=list(dec),
-                                        epoch=2000.0,
-                                        obs_metadata=self.obs_metadata,
-                                        camera=camera)
-
-        #test mismatches
-        self.assertRaises(RuntimeError, calculateFocalPlaneCoordinates, xPupil=numpy.array([xPupil[0]]), yPupil=yPupil,
-                          camera=camera)
-        self.assertRaises(RuntimeError, calculateFocalPlaneCoordinates, xPupil=xPupil, yPupil=numpy.array([yPupil[0]]),
-                          camera=camera)
-
-        self.assertRaises(RuntimeError, calculateFocalPlaneCoordinates, ra=numpy.array([ra[0]]), dec=dec,
-                                        epoch=2000.0, camera=camera)
-        self.assertRaises(RuntimeError, calculateFocalPlaneCoordinates, ra=ra, dec=numpy.array([dec[0]]),
-                                        epoch=2000.0,
-                                        obs_metadata=self.obs_metadata,
-                                        camera=camera)
-
-
-        ##########test calculatePixelCoordinates
-         #test that it actually runs
-        xx, yy = calculatePixelCoordinates(xPupil=xPupil, yPupil=yPupil, camera=camera)
-        xx, yy = calculatePixelCoordinates(ra=ra, dec=dec,
-                                                epoch=2000.0, obs_metadata=self.obs_metadata,
-                                                camera=camera)
-
-        #test without any coordinates
-        self.assertRaises(RuntimeError, calculatePixelCoordinates, camera=camera)
-
-        #test specifying both ra,dec and xPupil,yPupil
-        self.assertRaises(RuntimeError, calculatePixelCoordinates, ra=ra, dec=dec,
-                             xPupil=xPupil, yPupil=yPupil, camera=camera)
-
-        #test without camera
-        self.assertRaises(RuntimeError, calculatePixelCoordinates, xPupil=xPupil, yPupil=yPupil)
-        self.assertRaises(RuntimeError, calculatePixelCoordinates, ra=ra, dec=dec,
-                                        epoch=2000.0, obs_metadata=self.obs_metadata)
-
-        #test without epoch
-        self.assertRaises(RuntimeError, calculatePixelCoordinates, ra=ra, dec=dec,
-                                                obs_metadata=self.obs_metadata,
-                                                camera=camera)
-
-        #test without obs_metadata
-        self.assertRaises(RuntimeError, calculatePixelCoordinates, ra=ra, dec=dec,
-                                                epoch=2000.0,
-                                                camera=camera)
-
-        #test with lists
-        self.assertRaises(RuntimeError, calculatePixelCoordinates, xPupil=list(xPupil), yPupil=yPupil,
-                          camera=camera)
-        self.assertRaises(RuntimeError, calculatePixelCoordinates, xPupil=xPupil, yPupil=list(yPupil),
-                          camera=camera)
-
-        self.assertRaises(RuntimeError, calculatePixelCoordinates, ra=list(ra), dec=dec,
-                                        epoch=2000.0, camera=camera)
-        self.assertRaises(RuntimeError, calculatePixelCoordinates, ra=ra, dec=list(dec),
-                                        epoch=2000.0,
-                                        obs_metadata=self.obs_metadata,
-                                        camera=camera)
-
-        #test mismatches
-        self.assertRaises(RuntimeError, calculatePixelCoordinates, xPupil=numpy.array([xPupil[0]]), yPupil=yPupil,
-                          camera=camera)
-        self.assertRaises(RuntimeError, calculatePixelCoordinates, xPupil=xPupil, yPupil=numpy.array([yPupil[0]]),
-                          camera=camera)
-
-        self.assertRaises(RuntimeError, calculatePixelCoordinates, ra=numpy.array([ra[0]]), dec=dec,
-                                        epoch=2000.0, camera=camera)
-        self.assertRaises(RuntimeError, calculatePixelCoordinates, ra=ra, dec=numpy.array([dec[0]]),
-                                        epoch=2000.0,
-                                        obs_metadata=self.obs_metadata,
-                                        camera=camera)
-
-        chipNames = findChipName(xPupil=xPupil, yPupil=yPupil, camera=camera)
-        calculatePixelCoordinates(xPupil=xPupil, yPupil=yPupil, chipNames=chipNames, camera=camera)
-        self.assertRaises(RuntimeError, calculatePixelCoordinates, xPupil=xPupil, yPupil=yPupil,
-                                        camera=camera, chipNames=[chipNames[0]])
-
-        chipNames=findChipName(ra=ra, dec=dec, obs_metadata=self.obs_metadata, epoch=2000.0,
-                               camera=camera)
-        calculatePixelCoordinates(ra=ra, dec=dec, obs_metadata=self.obs_metadata, epoch=2000.0,
-                                  camera=camera, chipNames=chipNames)
-        self.assertRaises(RuntimeError, calculatePixelCoordinates, ra=ra, dec=dec, obs_metadata=self.obs_metadata,
-                          epoch=2000.0,
-                          camera=camera, chipNames=[chipNames[0]])
-
-
 
     def testApplyPrecession(self):
 
@@ -797,142 +579,6 @@ class astrometryUnitTest(unittest.TestCase):
         output=applyRefraction(0.25*numpy.pi,coeffs[0],coeffs[1])
 
         self.assertAlmostEqual(output,7.851689251070859132e-01,6)
-
-
-    def testPixelPos(self):
-        numpy.random.seed(42)
-
-        camera = camTestUtils.CameraWrapper().camera
-
-        nSamples = 100
-        ra = numpy.random.random_sample(nSamples)*radiansFromArcsec(100.0) + numpy.radians(self.obs_metadata.unrefractedRA)
-        dec = numpy.random.random_sample(nSamples)*radiansFromArcsec(100.0) + numpy.radians(self.obs_metadata.unrefractedDec)
-
-        pupilCoordinateArray = calculatePupilCoordinates(ra, dec, obs_metadata=self.obs_metadata,
-                                                         epoch=2000.0)
-
-        pixelCoordinateArray = calculatePixelCoordinates(ra=ra, dec=dec,
-                                                        obs_metadata=self.obs_metadata,
-                                                        epoch=2000.0, camera=camera)
-
-        chipNameList = findChipName(ra=ra, dec=dec,
-                                    obs_metadata=self.obs_metadata, epoch=2000.0, camera=camera)
-        self.assertTrue(numpy.all(numpy.isfinite(pupilCoordinateArray[0])))
-        self.assertTrue(numpy.all(numpy.isfinite(pupilCoordinateArray[1])))
-
-        for x, y, cname in zip(pixelCoordinateArray[0], pixelCoordinateArray[1],
-                               chipNameList):
-            if cname is None:
-                #make sure that x and y are not set if the object doesn't land on a chip
-                self.assertTrue(not numpy.isfinite(x) and not numpy.isfinite(y))
-            else:
-                #make sure the pixel positions are inside the detector bounding box.
-                self.assertTrue(afwGeom.Box2D(self.cat.camera[cname].getBBox()).contains(afwGeom.Point2D(x,y)))
-
-
-
-    def testPupilFromPixel(self):
-        """
-        Test the conversion between pixel coordinates and pupil coordinates
-        """
-        baseDir = os.path.join(getPackageDir('sims_coordUtils'),'tests','cameraData')
-        camera = ReturnCamera(baseDir)
-        epoch=2000.0
-        raCenter = 25.0
-        decCenter = -10.0
-        obs = ObservationMetaData(unrefractedRA=raCenter,
-                                  unrefractedDec=decCenter,
-                                  boundType='circle',
-                                  boundLength=0.1,
-                                  rotSkyPos=23.0,
-                                  mjd=52000.0)
-
-        raTrue, decTrue = observedFromICRS(numpy.array([numpy.radians(raCenter)]),
-                                           numpy.array([numpy.radians(decCenter)]),
-                                           obs_metadata=obs, epoch=epoch)
-
-        ra = []
-        dec = []
-
-        dx = 1.0e-4
-
-        for rr in numpy.arange(raTrue-20.0*dx, raTrue+20.0*dx, dx):
-            for dd in numpy.arange(decTrue-20.0*dx, decTrue+20.0*dx, dx):
-                ra.append(rr)
-                dec.append(dd)
-
-
-        ra = numpy.array(ra)
-        dec = numpy.array(dec)
-
-        xp, yp = calculatePupilCoordinates(ra, dec, obs_metadata=obs, epoch=2000.0)
-        chipNameList = findChipName(xPupil=xp, yPupil=yp, obs_metadata=obs, epoch=epoch,
-                                    camera=camera)
-        xPixList, yPixList = calculatePixelCoordinates(xPupil=xp, yPupil=yp, chipNames=chipNameList,
-                                               obs_metadata=obs, epoch=epoch, camera=camera)
-
-        xpTest, ypTest = pupilCoordinatesFromPixelCoordinates(xPixList, yPixList, chipNameList,
-                                                     camera=camera, obs_metadata=obs, epoch=epoch)
-
-        xpControl = numpy.array([xx if name is not None else numpy.NaN for (xx, name) in zip(xp, chipNameList)])
-        ypControl = numpy.array([yy if name is not None else numpy.NaN for (yy, name) in zip(yp, chipNameList)])
-
-        numpy.testing.assert_array_almost_equal(xpControl, xpTest, decimal=10)
-        numpy.testing.assert_array_almost_equal(ypControl, ypTest, decimal=10)
-
-
-
-    def testRaDecFromPixelCoordinates(self):
-        """
-        Test conversion from pixel coordinates to Ra, Dec
-        """
-
-        baseDir = os.path.join(getPackageDir('sims_coordUtils'),'tests','cameraData')
-        camera = ReturnCamera(baseDir)
-        epoch=2000.0
-
-        raCenter = 25.0
-        decCenter = -10.0
-        obs = ObservationMetaData(unrefractedRA=raCenter,
-                                  unrefractedDec=decCenter,
-                                  boundType='circle',
-                                  boundLength=0.1,
-                                  rotSkyPos=23.0,
-                                  mjd=52000.0)
-
-        raTrue, decTrue = observedFromICRS(numpy.array([numpy.radians(raCenter)]),
-                                           numpy.array([numpy.radians(decCenter)]),
-                                           obs_metadata=obs, epoch=epoch)
-
-        ra = []
-        dec = []
-
-        dx = 1.0e-4
-
-        for rr in numpy.arange(raTrue-20.0*dx, raTrue+20.0*dx, dx):
-            for dd in numpy.arange(decTrue-20.0*dx, decTrue+20.0*dx, dx):
-                ra.append(rr)
-                dec.append(dd)
-
-
-        ra = numpy.array(ra)
-        dec = numpy.array(dec)
-
-        chipNameList = findChipName(ra=ra, dec=dec, obs_metadata=obs, epoch=epoch, camera=camera)
-        pixelList = calculatePixelCoordinates(ra=ra, dec=dec, chipNames=chipNameList, obs_metadata=obs,
-                                           epoch=epoch, camera=camera)
-
-        raTest, decTest = raDecFromPixelCoordinates(pixelList[0], pixelList[1], chipNameList,
-                                                    obs_metadata=obs, epoch=epoch, camera=camera)
-
-
-
-        raControl = numpy.array([rr if name is not None else numpy.NaN for (rr, name) in zip(ra, chipNameList)])
-        decControl = numpy.array([dd if name is not None else numpy.NaN for (dd, name) in zip(dec, chipNameList)])
-
-        numpy.testing.assert_array_almost_equal(arcsecFromRadians(raControl), arcsecFromRadians(raTest), decimal=10)
-        numpy.testing.assert_array_almost_equal(arcsecFromRadians(decControl), arcsecFromRadians(decTest), decimal=10)
-
